Refactored smelly controller code by moving methods to both view and model
