+++ src/RiskController.java
@@ -1,9 +1,7 @@
 import javax.swing.*;
-import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.util.ArrayList;
-import java.util.HashMap;
 
 /**
  * The controller class handles all sort of events raised by the CountryButton countries
@@ -27,138 +25,6 @@ public class RiskController implements ActionListener{
         this.currentState = null;
     }
 
-    /**
-     * Takes in countries and returns a list of their corresponding country button instances from RiskView
-     * @author Harjap
-     * @param countries A list of all countries the player can use to attack
-     * @return A list of all the CountryButton instances
-     */
-    public ArrayList<CountryButton> convertCountryToCountryButtons(ArrayList<Country> countries){
-        // get hashmap from view that holds relationship between country name and country Button
-        HashMap<String, CountryButton> countryButtonHashMap = riskView.getCountryButtons();
-
-        // create empty arraylist for return
-        ArrayList<CountryButton> countryButtons = new ArrayList<CountryButton>();
-
-        for (Country c: countries){
-            countryButtons.add(countryButtonHashMap.get(c.getName()));
-        }
-
-        return countryButtons;
-    }
-
-    /**
-     * Gets the current player who is attacking
-     * @author Jason
-     * @return An instance of Player corresponding to the attacking player
-     */
-    private Player getAttackingPlayer() {
-        int turnIndex = riskModel.getTurnIndex();
-        ArrayList<Player> players = riskModel.getBoard().getPlayers();
-        return players.get(turnIndex);
-    }
-
-    /**
-     * Enables the attacking players CountryButton so that they may choose a country to attack from
-     * @author Jason
-     */
-    private void showAttackingCountries() {
-        // get all the Country Buttons that the current player owns
-        ArrayList<CountryButton> countryButtons = convertCountryToCountryButtons(getAttackingPlayer().getCountriesOwned());
-
-        for (CountryButton cb : countryButtons){
-            if (cb.getCountry().getPossibleAttacks().size() != 0 && cb.getCountry().getArmySize() >= 2) {
-                cb.setEnabled(true);
-            }
-        }
-        currentState = attackState.SHOW_DEFENDING_COUNTRIES;
-    }
-
-    /**
-     * Highlights all the countries the attacking player can attack from by creating a border around it
-     * @author Jason
-     * @param e An ActionEvent corresponding to the event created when a CountryButton was clicked
-     * @return The number of troops that the player chose to use in attacking the country
-     */
-    private int showDefendingCountries(ActionEvent e) {
-        CountryButton attackingCountry = (CountryButton) e.getSource();
-
-        // converts all countries that the attacking country can attack into their respective country button instance in the view and stores in a list
-        ArrayList<CountryButton> defendingCountries = convertCountryToCountryButtons(attackingCountry.getCountry().getAdjacentCountries());
-        defendingCountries.forEach(countryButton -> countryButton.setEnabled(true)); // enable all possible countries to attack
-        defendingCountries.forEach(countryButton -> countryButton.setBorder(BorderFactory.createLineBorder(Color.yellow, 3)));
-        // disables all attacking countries from being pressed
-        ArrayList<CountryButton> countryButtons = convertCountryToCountryButtons(getAttackingPlayer().getCountriesOwned());
-        countryButtons.forEach(countryButton -> countryButton.setEnabled(false));
-        countryButtons.forEach(countryButton -> countryButton.setBorder(null));
-
-        currentState = attackState.COMMENCE_ATTACK;
-
-        // asks for number of troops to attack with
-        return Integer.parseInt(getAttackingTroopCount(attackingCountry)); // number of troops to attack with
-    }
-
-    /**
-     * Creates a dialog box with a dropdown list asking the user for the number of troops they want to attack with
-     * @author Jason
-     * @param attackingCountry The country the player is attacking from
-     * @return The number of troops the player chose to attack with
-     */
-    private String getAttackingTroopCount(CountryButton attackingCountry) {
-        String[] options = {"OK"};
-        int troopCount = attackingCountry.getTroopCount()-1;
-        Integer[] troopList = buildTroopDropdownList(troopCount);
-        JPanel panel = new JPanel();
-        JLabel label = new JLabel("Select the number of troops to attack with: ");
-        JComboBox comboBox = new JComboBox(troopList);
-        comboBox.setSelectedIndex(0);
-        panel.add(label);
-        panel.add(comboBox);
-        int selectionObject = JOptionPane.showOptionDialog(riskView, panel, "Choose Troops", JOptionPane.NO_OPTION, JOptionPane.QUESTION_MESSAGE, null, options, options[0]);
-        String result;
-        while (selectionObject != 0) {
-            selectionObject = JOptionPane.showOptionDialog(riskView, panel, "Choose Troops", JOptionPane.NO_OPTION, JOptionPane.QUESTION_MESSAGE, null, options, options[0]);
-        }
-        result = comboBox.getSelectedItem().toString();
-        return result;
-    }
-
-    /**
-     * Create a list of Integers with the different amount of troops the player can use to attack with
-     * which ranges from 1..n-1
-     * @author Jason
-     * @param troopCount The number of troops available to attack with
-     * @return A Integer list containing the different options of troops the player can attack with
-     */
-    private Integer[] buildTroopDropdownList(int troopCount) {
-        Integer[] troopList = new Integer[troopCount];
-        // iterates one less to build a list from 1 - troopCount
-        for (int i = 0; i < troopList.length; i++) {
-            troopList[i] = troopCount;
-            troopCount--;
-        }
-        return troopList;
-    }
-
-    /**
-     * Handles the behaviour behind executing an attack by calling on the respective method in the model.
-     * This method also deals with setting up a continuous attack phase in case the user would like to attack again.
-     * @author Jason
-     * @param attackingCountry The country the user is attacking from
-     * @param defendingCountry The country the user is attacking
-     * @param troopCount The number of troops from the attacking country used to attack
-     */
-    private void performAttack(Country attackingCountry, Country defendingCountry, int troopCount) {
-        riskModel.attack(attackingCountry, defendingCountry, troopCount);
-
-        // disable all defending countries of the attacking country
-        ArrayList<CountryButton> countryButtons = convertCountryToCountryButtons(attackingCountry.getAdjacentCountries());
-        countryButtons.forEach(countryButton -> countryButton.setEnabled(false));
-        countryButtons.forEach(countryButton -> countryButton.setBorder(null));
-
-        showAttackingCountries(); // re-enable the user to choose countries for continuous attacking
-    }
-
     @Override
     public void actionPerformed(ActionEvent e) {
         if(e.getActionCommand().equals("country")) {
@@ -167,15 +33,18 @@ public class RiskController implements ActionListener{
             //identify what turn it is
             if (currentState == attackState.SHOW_DEFENDING_COUNTRIES) {
                 attackingCountry = countryButton.getCountry();
-                troopCount = showDefendingCountries(e);
+                troopCount = riskView.showDefendingCountries(e);
+                currentState = attackState.COMMENCE_ATTACK;
             } else if (currentState == attackState.COMMENCE_ATTACK) {
                 defendingCountry = countryButton.getCountry();
-                performAttack(attackingCountry,defendingCountry,troopCount);
+                riskModel.attack(attackingCountry, defendingCountry, troopCount);
+                riskView.performAttack(attackingCountry);
+                currentState = attackState.SHOW_DEFENDING_COUNTRIES;
             }
         }else if (e.getActionCommand().equals("endturn")){
             if (currentState == attackState.SHOW_DEFENDING_COUNTRIES || currentState == null) {
                 // disable all attacking countries of current player
-                ArrayList<CountryButton> countryButtons = convertCountryToCountryButtons(getAttackingPlayer().getCountriesOwned());
+                ArrayList<CountryButton> countryButtons = riskView.convertCountryToCountryButtons(riskModel.getAttackingPlayer().getCountriesOwned());
                 countryButtons.forEach(countryButton -> countryButton.setEnabled(false));
 
                 // moves to next player
@@ -188,8 +57,10 @@ public class RiskController implements ActionListener{
             currentState = attackState.SHOW_PLAYER_COUNTRIES;
             attack = (JButton) e.getSource();
             attack.setEnabled(false);
-            if (currentState == attackState.SHOW_PLAYER_COUNTRIES)
-                showAttackingCountries();
+            if (currentState == attackState.SHOW_PLAYER_COUNTRIES) {
+                riskView.showAttackingCountries();
+                currentState = attackState.SHOW_DEFENDING_COUNTRIES;
+            }
         }
 
     }
